==> notebooks/01_s3-dataAccess.ipynb <==
Test notebook to practice reading s3 files to SageMaker. The role returned by get_execution_role() needs to be added to the s3 Bucket policy within the Principal Key as "AWS":"arn:aws:iam::144273247162:role/SageMakerExecutionRole"

Currently the bucket is also allowing public access, so I'm not sure if this will work if the data is made private. However, I ran into an error when trying to access the public bucket without adding this role to the Policy.

## Imports

import os
import numpy as np
import pandas as pd
from matplotlib import pyplot as plt
import boto3
import sagemaker
from sagemaker import get_execution_role
import tempfile
import io

role = get_execution_role()
print(role)

region = boto3.Session().region_name
sess = sagemaker.Session()

bucket = 'cis5190-foodreviews'
file_path = 'Reviews_cleanText_noSW.csv'

s3_client = boto3.client('s3')
contents = s3_client.list_objects(Bucket=bucket)['Contents']

for f in contents:
    print(f['Key'])

obj = s3_client.get_object(Bucket=bucket, Key=file_path)

df = pd.read_csv(io.BytesIO(obj['Body'].read()))
df.head()

df.to_csv('Reviews_cleanText_noSW_sageMakerLocal.csv')
